add validations for workloads

+++ gridify/internal/deployer/deployer_test.go
@@ -60,7 +60,7 @@ func TestDeploy(t *testing.T) {
 
 		clientMock.
 			EXPECT().
-			GetAvailableNode(gomock.Any(), filter, uint64(Eco.Storage)).
+			GetAvailableNode(gomock.Any(), filter, Eco.Storage).
 			Return(uint32(0), errors.New("error"))
 
 		clientMock.
@@ -79,7 +79,7 @@ func TestDeploy(t *testing.T) {
 
 		clientMock.
 			EXPECT().
-			GetAvailableNode(gomock.Any(), filter, uint64(Eco.Storage)).
+			GetAvailableNode(gomock.Any(), filter, Eco.Storage).
 			Return(uint32(1), nil)
 
 		clientMock.
@@ -98,7 +98,7 @@ func TestDeploy(t *testing.T) {
 
 		clientMock.
 			EXPECT().
-			GetAvailableNode(gomock.Any(), filter, uint64(Eco.Storage)).
+			GetAvailableNode(gomock.Any(), filter, Eco.Storage).
 			Return(uint32(1), nil)
 
 		clientMock.
@@ -122,7 +122,7 @@ func TestDeploy(t *testing.T) {
 
 		clientMock.
 			EXPECT().
-			GetAvailableNode(gomock.Any(), filter, uint64(Eco.Storage)).
+			GetAvailableNode(gomock.Any(), filter, Eco.Storage).
 			Return(uint32(1), nil)
 
 		clientMock.
@@ -151,7 +151,7 @@ func TestDeploy(t *testing.T) {
 
 		clientMock.
 			EXPECT().
-			GetAvailableNode(gomock.Any(), filter, uint64(Eco.Storage)).
+			GetAvailableNode(gomock.Any(), filter, Eco.Storage).
 			Return(uint32(1), nil)
 
 		clientMock.
@@ -185,7 +185,7 @@ func TestDeploy(t *testing.T) {
 
 		clientMock.
 			EXPECT().
-			GetAvailableNode(gomock.Any(), filter, uint64(Eco.Storage)).
+			GetAvailableNode(gomock.Any(), filter, Eco.Storage).
 			Return(uint32(1), nil)
 
 		clientMock.
@@ -224,7 +224,7 @@ func TestDeploy(t *testing.T) {
 
 		clientMock.
 			EXPECT().
-			GetAvailableNode(gomock.Any(), filter, uint64(Eco.Storage)).
+			GetAvailableNode(gomock.Any(), filter, Eco.Storage).
 			Return(uint32(1), nil)
 
 		clientMock.
@@ -264,7 +264,7 @@ func TestDeploy(t *testing.T) {
 
 		clientMock.
 			EXPECT().
-			GetAvailableNode(gomock.Any(), filter, uint64(Eco.Storage)).
+			GetAvailableNode(gomock.Any(), filter, Eco.Storage).
 			Return(uint32(1), nil)
 
 		clientMock.
